Bug fixed, changed int pointer to double pointer.

+++ src/matrix.c
@@ -46,10 +46,10 @@ Matrix* matrix_identity(int w,int h)
     return output;
 }
 //merge sort
-static void merge_aux(int *a, int n, int m)
+static void merge_aux(double *a, int n, int m)
 {
 	int i, j, k;
-	int *x = malloc(n * sizeof(int));
+	double *x = malloc(n * sizeof(double));
 	for (i = 0, j = m, k = 0; k < n; k++) 
 	{
 		x[k] = j == n ? a[i++]
@@ -63,7 +63,7 @@ static void merge_aux(int *a, int n, int m)
 	}
 	free(x);
 }
-static void merge_sort_aux(int *a, int n)
+static void merge_sort_aux(double *a, int n)
 {
 	if (n < 2) return;
 	int m = n / 2;
